Remove unnecessary rescue ActiveRecord::RecordNotFound
The database foreign key constraint between the sessions and users
tables prevents sessions from existing with a user_id that points to a
nonexistent user. As a result, session_from_jwt.user will never raise
ActiveRecord::RecordNotFound, and attempting to authenticate with a non-
existent session will result in an ActiveRecord::RecordNotFound that is
already rescued from by re-raising an HTTP::Errors::UnauthorizedError.